requirements.txt
dill==0.3.6
hydra-core==1.3.2
lightgbm==3.3.3
mapie==0.5.0
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
omegaconf==2.3.0
pandas==1.3.5
python_igraph==0.10.2
pytorch_lightning==1.8.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.13.0
tqdm==4.57.0
